pyproject.toml
[project]
name = "motorcycle_coach"
version = "0.1.0"
dependencies = [
    "openenv-core>=0.2.0",
    "openai>=1.0.0",
    "uvicorn>=0.24.0",
    "pydantic>=2.0.0",
    "fastapi>=0.104.0",
]
requires-python = ">=3.10"

[project.scripts]
server = "server.app:main"

[build-system]
requires = ["setuptools>=61.0"]
build-backend = "setuptools.build_meta"